Update installing.md
Hi not sure if this is right, just signed up for GitHub after using Jellyfin for a few weeks. Found the Debian repository instructions tricky to follow as things were marked as code which I don’t think need to be for legibility to non-tech users (like me). Thought I’d try and start to give back something to the project as I can’t code. 

Hope this is ok.