update out-of-date information about rclone/cache (#342)
rclone cache hasn't been required to prevent API bans for some time since the chunker was introduced a while ago.  The cache backend is outdated and not maintained.  Mergerfs isn't used to prevent API bans either.  It is used to share a local and cloud file system as an overlay.