Kodi native mode with mounted shares
Mounted shares can be referenced in the library and can be beneficial
from a performance standpoint since it’s offloaded to the OS directly.
This is a popular approach on “low end” Kodi clients, like Raspberry Pi
for instance.

The downside of using mounts is that the mount point is not
interoperable between different operating systems.

How to actually mount a share is way outside the scope of this
documentation, but I added some examples from different operating
systems.