Updated the podman section in installing.md  (#500)
* Update installing.md

Updated the podman section, so the container is a little less exposed by default. I.e. `--privileged` is removed and only the ports needed are passed. 

Then I added a subsection on managing the service with systemd. Along with some sandboxing options that have been tested to work, since that can involve a lot of guess work.