Add information about comics
Add documentation about which comic metadata providers are currently
supported. This includes how to best name the comics that a future
online metadata provider can find the correct comic for most use cases.

The naming scheme is inspired by "comictagger" but simpliefies how
comic volumes are handled. The reason for this is that in the manga
world, an issue is referred to as a volume and ComicVine refers to a
series as a volume.

Instead of having to define what a volume is, the issue and series names
as well as the release year are used in the name. This should allow
a potential online metadata provider to find most comics.