Update troubleshooting.md
How to unlock Blocked User

When the admin account is locked out you have only the way over the jellyfin.db with sqlite query to unlock the admin user. At the moment there i no way documented on the internet. The corresponding issue is these:
https://github.com/jellyfin/jellyfin/issues/6645#event-5412515320
after a few times back and forth cvium and i switched to the matrix chat and got the problem fixed there

Co-Authored-By: Tim Gels <43609220+TimGels@users.noreply.github.com>
Co-Authored-By: Lukáš Kucharczyk <lukas@kucharczyk.xyz>