Document removal of the time limit before saving
This documents the change in Jellyfin-web, that books will be saved
more accurately and are no longer saved in a time-interval.

This is done by removing the information entirely as the user would
expect the reading progress to be saved on user-action. Thus it is not
necessary to explicitly state this behaviour.