Created an Addons section
I originally added a github repo stars in plugins. I decided to add an addons repo with all the possible addons not already listed in jellyfins docs. I figured this was best since it didnt fit any of the other categories and to make it simpler for people to find what they want in the least amount of clicks.